Load ordering issues. Much pfaffing about.
The way emacs loads and configures itself means that settings near the
top of your initialization stack can be clobbered by settings brought
in later on as part of packages. Anyhow, I started moving things
around, and then it got... weird. It's still weird, but at least not
fully hideous.