Load Sly from a local source checkout instead of melpa.
this is for various debugging reasons. Namely, the REPL shortcuts weren't being activated in the REPL when the comma was entered. It turned out that something in the dependency graph of the system that I was developing was clobbering the REPL in a way I haven't yet ascertained. If you're using the system :CLOG, be aware.